extract localizedRoutePath and localizedRouteConfig tests
