+++ core/lib/multistore/test/unit/localizedRouteConfig.spec.ts
@@ -1,0 +1,83 @@
+import {
+  localizedRouteConfig
+} from '@vue-storefront/core/lib/multistore'
+
+jest.mock('@vue-storefront/core/app', () => ({
+  createApp: jest.fn(),
+  router: {
+    addRoutes: jest.fn()
+  }
+}))
+jest.mock('@vue-storefront/i18n', () => ({ loadLanguageAsync: jest.fn() }))
+jest.mock('@vue-storefront/core/lib/logger', () => ({
+  Logger: {}
+}))
+jest.mock('config', () => ({}))
+
+describe('localizedRouteConfig', () => {
+  it('create new route object with storeCode', () => {
+    const storeCode = 'de'
+    const route = {
+      path: '/test',
+      name: 'test'
+    }
+    const expectedRoute = {
+      path: '/de/test',
+      name: 'de-test'
+    }
+
+    expect(localizedRouteConfig(route, storeCode)).toEqual(expectedRoute)
+  })
+
+  it('change only route name for child route', () => {
+    const storeCode = 'de'
+    const childRoute = {
+      path: '/test2',
+      name: 'test2'
+    }
+    const expectedRoute = {
+      path: '/test2',
+      name: 'de-test2'
+    }
+
+    expect(localizedRouteConfig(childRoute, storeCode, true)).toEqual(expectedRoute)
+  })
+
+  it('add localization for nested routes', () => {
+    const storeCode = 'de'
+    const route = {
+      path: '/test',
+      name: 'test',
+      children: [
+        {
+          path: 'test2',
+          name: 'test2',
+          children: [
+            {
+              path: '/test3',
+              name: 'test3'
+            }
+          ]
+        }
+      ]
+    }
+    const expectedRoute = {
+      path: '/de/test',
+      name: 'de-test',
+      children: [
+        {
+          path: 'test2',
+          name: 'de-test2',
+          children: [
+            {
+              path: '/test3',
+              name: 'de-test3'
+            }
+          ]
+        }
+      ]
+    }
+
+    expect(localizedRouteConfig(route, storeCode)).toEqual(expectedRoute)
+  })
+})
